requirements.txt
streamlit
PyPDF2
matplotlib
google-generativeai